.Trash/backend 14-59-45-783/requirements.txt
# Core dependencies
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0

# LangChain
langchain==1.2.7
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-naver==0.1.1
langchain-community==0.4.1

# OpenAI
openai==1.109.1

# Google
google-genai==1.60.0

# HTTP
httpx==0.28.1
aiohttp==3.13.3
requests==2.32.5

# Vector DB
pymilvus==2.6.6

# Data & Utils
python-dateutil==2.9.0.post0
typing-extensions==4.15.0
python-multipart==0.0.22

pymongo
pymysql
polars
motor
boto3

kiwipiepy
